CPMU-3 Implement partition table support in the utility library
new partition table format based on FlatTables.
