+++ build.gradle
@@ -1,6 +1,9 @@
+import io.netifi.flatbuffers.plugin.tasks.FlatBuffers
+
 plugins {
     id 'java'
     id 'maven-publish'
+    id "io.netifi.flatbuffers" version "1.0.7"
 }
 
 group 'net.uberfoo.cpm'
@@ -13,6 +16,7 @@ repositories {
 dependencies {
     implementation 'org.jetbrains:annotations:23.0.0'
     implementation 'org.slf4j:slf4j-api:2.0.6'
+    implementation 'com.google.flatbuffers:flatbuffers-java:23.1.21'
 
     testImplementation(platform('org.apache.logging.log4j:log4j-bom:2.19.0'))
 
@@ -25,12 +29,20 @@ dependencies {
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testImplementation 'org.hamcrest:hamcrest:2.2'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
 }
 
 java {
     withJavadocJar()
     withSourcesJar()
+}
+
+flatbuffers {
+    language = 'java'
+}
+
+task createFlatBuffers(type: FlatBuffers) {
+    outputDir = file("build/generated/sources/flatbuffers")
 }
 
 test {
