--- src/main/java/module-info.java
@@ -1,9 +1,0 @@
-/**
- * CP/M utilities module
- */
-module net.uberfoo.cpm.cpmutils {
-    requires org.slf4j;
-    requires org.jetbrains.annotations;
-
-    exports net.uberfoo.cpm.filesystem;
-}
